requirements.txt
absl-py==2.0.0
astunparse==1.6.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.58.5
gast==0.4.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
grpcio==1.48.2
gunicorn==20.1.0
h5py==3.14.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
keras==2.15.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
ml-dtypes==0.2.0
namex==0.1.0
numpy==1.24.4
oauthlib==3.3.1
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
Pillow==9.0.0
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.14.1
zipp==3.23.0
waitress==2.1.2